requirements.txt
dash==2.17.0
plotly==5.21.0
pandas==2.2.2
notion-client==2.2.1
python-dotenv==1.0.1
ipykernel==6.29.4
gunicorn==21.2.0
